sh函数封装安装过程

+++ script/pipsetup.sh
@@ -1,14 +1,30 @@
-echo '...you will  steup  python pip...'
-pipversiondefault=1.5.5
-yum  install   python-setuptools
+#check if the supervisor has been setup
+installpip()
+{
+    echo '...you will  steup  python pip...'
+    pipversiondefault=1.5.5
+    yum install python-setuptools
+    wget --no-check-certificate https://github.com/pypa/pip/archive/$pipversiondefault.tar.gz
+    tar zvxf $pipversiondefault.tar.gz    
+    cd pip-$pipversiondefault/
+    python setup.py install
+}
+installsupervisor()
+{
+    echo  "installing ...."
+    echo '...You will  steup  supervisor...'
+    file="supervisor.conf"
+    cd /etc
+    if [  -f "$file" ]; then
+        echo "supervisor has been  setup,exiting..."
+        exit
+    else
+        pip install supervisor
+        echo_supervisord_conf >/etc/supervisor.conf
+    fi
+}
+installpip
+installsupervisor
 
-wget --no-check-certificate https://github.com/pypa/pip/archive/$pipversiondefault.tar.gz
 
-tar zvxf $pipversiondefault.tar.gz    #解压文件
-cd pip-$pipversiondefault/
-python setup.py install
-
-
-echo '...you will  steup  supervisor...'
-pip  install  supervisor
-echo_supervisord_conf >/etc/supervisor.conf 
+
